+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -45,6 +45,8 @@ CHANGELOG
    `Symfony\Component\Validator\DependencyInjection\AddValidatorInitializersPass` instead
  * Deprecated `AddConstraintValidatorsPass`, use
    `Symfony\Component\Validator\DependencyInjection\AddConstraintValidatorsPass` instead
+ * Deprecated `ValidateWorkflowsPass`, use 
+   `Symfony\Component\Workflow\DependencyInjection\ValidateWorkflowsPass` instead
 
 3.2.0
 -----